+++ parsley-core/src/ghc/Parsley/Internal/Backend/Machine/Instructions.hs
@@ -24,7 +24,7 @@ module Parsley.Internal.Backend.Machine.Instructions (
     -- * Smart Instructions
     _App, _Fmap, _Modify, _Make, _Put, _Get,
     -- * Smart Meta-Instructions
-    addCoins, refundCoins, drainCoins, giveBursary, prefetchChar
+    addCoins, refundCoins, drainCoins, giveBursary, prefetchChar, blockCoins
   ) where
 
 import Data.Kind                                    (Type)
@@ -260,6 +260,12 @@ data MetaInstr (n :: Nat) where
 
   @since 1.5.0.0 -}
   PrefetchChar :: Bool -> MetaInstr (Succ n)
+  {-|
+  True meta instruction: does /nothing/ except for reset coin count during coin analysis.
+
+  @since 1.6.0.0
+  -}
+  BlockCoins :: MetaInstr n
 
 mkCoin :: (Coins -> MetaInstr n) -> Coins -> Fix4 (Instr o) xs n r a -> Fix4 (Instr o) xs n r a
 mkCoin _    (Coins 0 0) = id
@@ -271,7 +277,8 @@ Smart-constuctor around `AddCoins`.
 @since 1.5.0.0
 -}
 addCoins :: Coins -> Fix4 (Instr o) xs (Succ n) r a -> Fix4 (Instr o) xs (Succ n) r a
-addCoins = mkCoin AddCoins
+addCoins (Coins 1 1) = id
+addCoins coins       = mkCoin AddCoins coins
 
 {-|
 Smart-constuctor around `RefundCoins`.
@@ -300,10 +307,18 @@ giveBursary = mkCoin GiveBursary
 {-|
 Smart-constructor around `PrefetchChar`.
 
-@since 1.5.0.0 
+@since 1.5.0.0
 -}
 prefetchChar :: Bool -> Fix4 (Instr o) xs (Succ n) r a -> Fix4 (Instr o) xs (Succ n) r a
 prefetchChar check = In4 . MetaInstr (PrefetchChar check)
+
+{-|
+Smart-constructor around `PrefetchChar`.
+
+@since 1.6.0.0
+-}
+blockCoins :: Fix4 (Instr o) xs (Succ n) r a -> Fix4 (Instr o) xs (Succ n) r a
+blockCoins = In4 . MetaInstr BlockCoins
 
 {-|
 Applies a value on the top of the stack to a function on the second-most top of the stack.
@@ -389,31 +404,32 @@ instance Show (Fix4 (Instr o) xs n r a) where
   show = ($ "") . getConst4 . cata4 (Const4 . alg)
     where
       alg :: forall xs n r a. Instr o (Const4 (String -> String)) xs n r a -> String -> String
-      alg Ret                 = "Ret"
-      alg (Call μ k)          = "(Call " . shows μ . " " . getConst4 k . ")"
-      alg (Jump μ)            = "(Jump " . shows μ . ")"
-      alg (Push x k)          = "(Push " . shows x . " " . getConst4 k . ")"
-      alg (Pop k)             = "(Pop " . getConst4 k . ")"
-      alg (Lift2 f k)         = "(Lift2 " . shows f . " " . getConst4 k . ")"
-      alg (Sat f k)           = "(Sat " . shows f . " " . getConst4 k . ")"
-      alg Empt                = "Empt"
-      alg (Commit k)          = "(Commit " . getConst4 k . ")"
-      alg (Catch p h)         = "(Catch " . getConst4 p . " " . shows h . ")"
-      alg (Tell k)            = "(Tell " . getConst4 k . ")"
-      alg (Seek k)            = "(Seek " . getConst4 k . ")"
-      alg (Case p q)          = "(Case " . getConst4 p . " " . getConst4 q . ")"
-      alg (Choices fs ks def) = "(Choices " . shows fs . " [" . intercalateDiff ", " (map getConst4 ks) . "] " . getConst4 def . ")"
-      alg (Iter μ l h)        = "{Iter " . shows μ . " " . getConst4 l . " " . shows h . "}"
-      alg (Join φ)            = shows φ
-      alg (MkJoin φ p k)      = "(let " . shows φ . " = " . getConst4 p . " in " . getConst4 k . ")"
-      alg (Swap k)            = "(Swap " . getConst4 k . ")"
-      alg (Dup k)             = "(Dup " . getConst4 k . ")"
-      alg (Make σ a k)        = "(Make " . shows σ . " " . shows a . " " . getConst4 k . ")"
-      alg (Get σ a k)         = "(Get " . shows σ . " " . shows a . " " . getConst4 k . ")"
-      alg (Put σ a k)         = "(Put " . shows σ . " " . shows a . " " . getConst4 k . ")"
-      alg (LogEnter _ k)      = getConst4 k
-      alg (LogExit _ k)       = getConst4 k
-      alg (MetaInstr m k)     = "[" . shows m . "] " . getConst4 k
+      alg Ret                      = "Ret"
+      alg (Call μ k)               = "(Call " . shows μ . " " . getConst4 k . ")"
+      alg (Jump μ)                 = "(Jump " . shows μ . ")"
+      alg (Push x k)               = "(Push " . shows x . " " . getConst4 k . ")"
+      alg (Pop k)                  = "(Pop " . getConst4 k . ")"
+      alg (Lift2 f k)              = "(Lift2 " . shows f . " " . getConst4 k . ")"
+      alg (Sat f k)                = "(Sat " . shows f . " " . getConst4 k . ")"
+      alg Empt                     = "Empt"
+      alg (Commit k)               = "(Commit " . getConst4 k . ")"
+      alg (Catch p h)              = "(Catch " . getConst4 p . " " . shows h . ")"
+      alg (Tell k)                 = "(Tell " . getConst4 k . ")"
+      alg (Seek k)                 = "(Seek " . getConst4 k . ")"
+      alg (Case p q)               = "(Case " . getConst4 p . " " . getConst4 q . ")"
+      alg (Choices fs ks def)      = "(Choices " . shows fs . " [" . intercalateDiff ", " (map getConst4 ks) . "] " . getConst4 def . ")"
+      alg (Iter μ l h)             = "{Iter " . shows μ . " " . getConst4 l . " " . shows h . "}"
+      alg (Join φ)                 = shows φ
+      alg (MkJoin φ p k)           = "(let " . shows φ . " = " . getConst4 p . " in " . getConst4 k . ")"
+      alg (Swap k)                 = "(Swap " . getConst4 k . ")"
+      alg (Dup k)                  = "(Dup " . getConst4 k . ")"
+      alg (Make σ a k)             = "(Make " . shows σ . " " . shows a . " " . getConst4 k . ")"
+      alg (Get σ a k)              = "(Get " . shows σ . " " . shows a . " " . getConst4 k . ")"
+      alg (Put σ a k)              = "(Put " . shows σ . " " . shows a . " " . getConst4 k . ")"
+      alg (LogEnter _ k)           = getConst4 k
+      alg (LogExit _ k)            = getConst4 k
+      alg (MetaInstr BlockCoins k) = getConst4 k
+      alg (MetaInstr m k)          = "[" . shows m . "] " . getConst4 k
 
 instance Show (Handler o (Const4 (String -> String)) (o : xs) n r a) where
   show (Same yes no) = "(Dup (Tell (Lift2 same (If " (getConst4 yes (" " (getConst4 no "))))")))
@@ -425,3 +441,4 @@ instance Show (MetaInstr n) where
   show (DrainCoins n)   = "Using " ++ show n ++ " coins"
   show (GiveBursary n)  = "Bursary of " ++ show n ++ " coins"
   show (PrefetchChar b) = "Prefetch character " ++ (if b then "with" else "without") ++ " length-check"
+  show BlockCoins       = ""