+++ parsley-core/test/RegressionTest.hs
@@ -1,0 +1,11 @@
+module Main where
+
+import Test.Tasty
+import qualified Regression.Issue27 as Issue27
+
+main :: IO ()
+main = defaultMain tests
+
+tests :: TestTree
+tests = testGroup "Regression Tests" [ Issue27.tests
+                                     ]